# server/docs/postman-openapi-full.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zalo Clone — Chat Server API (full)",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://unlimited-feedback-crossing-academics.trycloudflare.com",
      "description": "Qua Cloudflare tunnel (gọi từ máy khác)"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local trực tiếp tại máy chủ"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "AGENT_API_KEY hoặc ADMIN_API_KEY trong `server/.env`. GET /api/media/* được miễn."
      }
    },
    "parameters": {
      "TypeParam": {
        "name": "type",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "enum": [
            0,
            1
          ]
        },
        "description": "Loại thread (zca-js ThreadType): **0** = người dùng (chat 1-1), **1** = nhóm."
      },
      "ThreadIdParam": {
        "name": "threadId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID hội thoại: uid người dùng hoặc groupId."
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Thiếu hoặc sai API key (header `X-Api-Key`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mô tả lỗi (tiếng Việt)."
          },
          "code": {
            "type": "string",
            "nullable": true,
            "description": "Mã lỗi gốc từ Zalo (nếu có)."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Passthrough": {
        "type": "object",
        "additionalProperties": true,
        "description": "Object thô trả về từ Zalo (zca-js). Cấu trúc tuỳ thao tác."
      },
...
    }
  },
  "paths": {
...
    "/api/threads/{type}/{threadId}/polls": {
      "post": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Tạo bình chọn",
        "parameters": [
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/ThreadIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Tham số poll (question, options…) theo zca-js."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      },
      "get": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Danh sách bình chọn của hội thoại",
        "parameters": [
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/ThreadIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/polls/{pollId}": {
      "get": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Chi tiết bình chọn",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/polls/{pollId}/lock": {
      "post": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Khoá bình chọn",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/polls/{pollId}/options": {
      "post": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Thêm lựa chọn",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "options"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/polls/{pollId}/vote": {
      "post": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Bình chọn",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "optionId": {
                    "description": "ID lựa chọn (hoặc mảng)."
                  }
                },
                "required": [
                  "optionId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/polls/{pollId}/share": {
      "post": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Chia sẻ bình chọn",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/threads/{type}/{threadId}/reminders": {
      "post": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Tạo nhắc hẹn",
        "parameters": [
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/ThreadIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Tham số reminder (title, time…) theo zca-js."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      },
      "get": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Danh sách nhắc hẹn của hội thoại",
        "parameters": [
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/ThreadIdParam"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/threads/{type}/{threadId}/reminders/{reminderId}": {
      "patch": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Sửa nhắc hẹn",
        "parameters": [
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/ThreadIdParam"
          },
          {
            "name": "reminderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      },
      "delete": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Xoá nhắc hẹn",
        "parameters": [
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/ThreadIdParam"
          },
          {
            "name": "reminderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/reminders/{reminderId}": {
      "get": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Chi tiết nhắc hẹn",
        "parameters": [
          {
            "name": "reminderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/reminders/{reminderId}/responses": {
      "get": {
        "tags": [
          "Bình chọn & Nhắc hẹn"
        ],
        "summary": "Phản hồi của nhắc hẹn",
        "parameters": [
          {
            "name": "reminderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    }
  }
}